Create Obstacle Avoiding Robot (Arduino | Bluetooth)
This is the first project under the subject of Robotic Application Development and IOT in my undergraduate degree program. This robot is capable of  not hitting the walls in front of it even we give the command to go forward. When it detects a obstacle in front, it automatically go backward and scan left and right, then turn to a side which has more space than previous reading.